[Refactoring] move functions about DB to db_api.js
